requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch>=1.11.0
torchvision>=0.12.0
segmentation-models-pytorch==0.3.3
albumentations==1.3.1
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3